requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.39.1
django-cors-headers==2.4.0
django-extensions==2.1.5
django-js-asset==1.1.0
django-mptt==0.9.1
django-pyodbc-azure==2.1.0.0
django-rest-auth==0.9.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.0.1
PyJWT==1.7.1
pyodbc==4.0.25
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
